Add log for flowgraph
